Project gfloat, and say what its width costs
37 GTK members take or return a gfloat, and most are ordinary property
pairs — Label.xalign, Entry.alignment, Frame.labelAlign,
AspectFrame.ratio. All of them were refused, on the grounds that
ScriptC's float slice is binary64 and a 32-bit float would silently
widen every value it carried.

That was right about the language and wrong about the ABI. A float in a
foreign signature is a slot, not a second precision to compute in, so
gfloat declares the JavaScript-number conversion like every other
numeric type here and the pinned compiler admits the scalar in a slot
and nowhere else.

What makes this one different is that its crossing is not exact, and
the honest thing is to say so rather than to refuse the type. Reading a
gfloat is lossless — every float is a double. Writing one rounds to
nearest float, which is what storing a double in 32 bits means and what
gtk_label_set_xalign has always done; a magnitude past the float range
becomes an infinity as IEEE says, rather than as C's undefined
conversion would leave to the target. There is no checked alternative
worth having: it would reject most decimals a caller can write.

The widget gate states the consequence instead of hiding it.
`heading.xalign = 0.25` round-trips untouched because 0.25 is a float;
`heading.yalign = 0.1` comes back different from 0.1 and within
binary32's tolerance of it. Both assertions run in a real window
through both backends, and flipping the first one fails the gate.

+++ tests/gobject-adapter.test.ts
@@ -481,10 +481,10 @@ test("unsupported GObject signal shapes fail with a stable diagnostic", () => {
 });
 
 test("unsupported GObject signal payloads fail at the exact parameter", () => {
-  /* `gfloat` is refused everywhere it appears: ScriptC's float slice is exactly
-   * f64, so admitting a 32-bit payload would widen every value it carries. The
-   * point here is that the refusal names the parameter rather than the
-   * signal. */
+  /* `gsize` is absent from the scalar table on purpose — its width should
+   * come from probe evidence rather than from a table that assumes an ABI —
+   * so a payload carrying one is outside the slice. The point here is that
+   * the refusal names the parameter rather than the signal. */
   const source = girSource.replace(
     `<return-value transfer-ownership="none">
           <type name="none" c:type="void"/>
@@ -495,7 +495,7 @@ test("unsupported GObject signal payloads fail at the exact parameter", () => {
         </return-value>
         <parameters>
           <parameter name="fraction" transfer-ownership="none">
-            <type name="gfloat" c:type="gfloat"/>
+            <type name="gsize" c:type="gsize"/>
           </parameter>
         </parameters>
       </glib:signal>`,
